footwear_rag/requirements.txt
# Footwear RAG Agent - Python Dependencies

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Snowflake
snowflake-connector-python>=3.5.0
snowflake-sqlalchemy>=1.5.0

# Data Validation
pydantic>=2.5.0

# File Uploads
python-multipart>=0.0.6

# Image Processing
Pillow>=10.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0  # For reading Excel files (.xlsx)

# Configuration
PyYAML>=6.0
python-dotenv>=1.0.0

# HTTP Client (for testing)
httpx>=0.25.0

# Development (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
